requirements.txt
Babel==2.10.3
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.36.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
passlib==1.7.4
Pillow==9.2.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tinycss2==1.1.1
tkcalendar==1.6.1
urllib3==1.26.11
webencodings==0.5.1
